Factor out p2p deployer.
